crypto-bot-delivery/src/delivery_api/requirements.txt
fastapi
uvicorn[standard] # For running the server
pydantic
python-jose[cryptography]
passlib[bcrypt]
